Enable -Wconversion in window.c

+++ src/nvim/eval.c
@@ -1826,9 +1826,9 @@ ex_let_one (
                       && vim_strchr(endchars, *skipwhite(p)) == NULL))
       EMSG(_(e_letunexp));
     else {
-      long n;
+      int n;
       int opt_type;
-      long numval;
+      int numval;
       char_u      *stringval = NULL;
       char_u      *s;
 
@@ -4538,7 +4538,7 @@ get_option_tv (
 )
 {
   char_u      *option_end;
-  long numval;
+  int numval;
   char_u      *stringval;
   int opt_type;
   int c;
