+++ src/nvim/normal.h
@@ -50,9 +50,9 @@ typedef struct cmdarg_S {
   int ncharC1;                  /* first composing character (optional) */
   int ncharC2;                  /* second composing character (optional) */
   int extra_char;               /* yet another character (optional) */
-  long opcount;                 /* count before an operator */
-  long count0;                  /* count before command, default 0 */
-  long count1;                  /* count before command, default 1 */
+  int opcount;                  /* count before an operator */
+  int count0;                   /* count before command, default 0 */
+  int count1;                  /* count before command, default 1 */
   int arg;                      /* extra argument from nv_cmds[] */
   int retval;                   /* return: CA_* values */
   char_u      *searchbuf;       /* return: pointer to search pattern or NULL */
